[수정] RacingView 를 InputView 와 OutputView 로 분리
