+++ src/main/java/racing/Main.java
@@ -16,7 +16,8 @@ import java.util.List;
 
 import racing.domain.Car;
 import racing.domain.RacingGame;
-import racing.view.RacingView;
+import racing.view.InputView;
+import racing.view.OutputView;
 
 /**
  * 레이싱 게임을 진행하는 컨트롤러 클래스
@@ -24,33 +25,33 @@ import racing.view.RacingView;
  * @author imkimheejoo
  * @author MrKwon
  * @version 2.0.0
- * @see RacingView
+ * @see InputView
  * @see RacingGame
  */
 public class Main {
     public static void main(String[] args) {
-        RacingView racingView = new RacingView();
-        List<Car> cars = getCars(racingView);
-        int runs = getRuns(racingView);
-        runRacingGame(racingView, runs, new RacingGame(cars));
+        InputView inputView = new InputView();
+        List<Car> cars = getCars(inputView);
+        int runs = getRuns(inputView);
+        runRacingGame(runs, new RacingGame(cars));
     }
 
-    private static List<Car> getCars(RacingView racingView) {
-        racingView.printMessageCarNames();
-        return racingView.getCarNames();
+    private static List<Car> getCars(InputView inputView) {
+        OutputView.printMessageCarNames();
+        return inputView.getCarNames();
     }
 
-    private static int getRuns(RacingView racingView) {
-        racingView.printMessageRuns();
-        return racingView.getRuns();
+    private static int getRuns(InputView inputView) {
+        OutputView.printMessageRuns();
+        return inputView.getRuns();
     }
 
-    private static void runRacingGame(RacingView racingView, int runs, RacingGame racingGame) {
-        racingView.printResultTitle();
+    private static void runRacingGame(int runs, RacingGame racingGame) {
+        OutputView.printResultTitle();
         for (int i = 0; i < runs; i++) {
             racingGame.runEachGame();
-            racingView.printEachGameResult(racingGame);
+            OutputView.printEachGameResult(racingGame);
         }
-        racingView.printResultTotal(racingGame);
+        OutputView.printResultTotal(racingGame);
     }
 }
